Restructure and add lint
