+++ package.json
@@ -2,18 +2,20 @@
   "name": "homebridge-alexa",
   "version": "0.8.0",
   "description": "Control your Homebridge devices with Amazon Alexa.",
-  "main": "plugin.js",
+  "main": "src/plugin.js",
   "scripts": {
     "test": "jest --forceExit --detectOpenHandles --verbose=true --silent ",
     "api": "documentation build plugin.js lib/alexaActions.js lib/alexaLocal.js lib/alexaMessages.js lib/alexaTranslator.js -f md --config docs/documentation.yml > docs/API.md",
     "document": "./gh-md-toc --hide-footer README.md ;./gh-md-toc --hide-footer Troubleshooting.md ;./gh-md-toc --hide-footer Installation.md; rm README.md.orig.* README.md.toc.* Troubleshooting.md.orig.* Troubleshooting.md.toc.* Installation.md.orig.* Installation.md.toc.*",
-    "watch": "nodemon"
+    "watch": "nodemon",
+    "lint": "eslint --max-warnings=10 .",
+    "lint:fix": "eslint --fix --max-warnings=0 ."
   },
   "nodemonConfig": {
     "watch": [],
     "ext": "js",
     "ignore": [],
-    "exec": "~/npm/bin/homebridge -D -P ~/Code",
+    "exec": "~/npm/bin/homebridge -U ./test/hbConfig -T -D -P .",
     "signal": "SIGTERM",
     "env": {
       "NODE_OPTIONS": "--trace-warnings"
@@ -33,27 +35,27 @@
   ],
   "devDependencies": {
     "@oznu/hap-client": "^1.9.0",
-    "@types/jest": "^29.5.10",
-    "@types/node": "^22.7.0",
+    "@types/node": "^22.10.1",
+    "@types/jest": "^29.5.14",
+    "@eslint/js": "^9.14.0",
+    "eslint": "^8.57.1",
+    "eslint-plugin-format": "^0.1.2",
+    "eslint-plugin-jest": "^28.8.3",
     "babel-jest": "^29.7.0",
+    "globals": "^15.12.0",
     "jest": "^29.7.0",
     "mqtt-connection": "^4.1.0",
     "nodemon": "^3.1.4",
     "rimraf": "^6.0.1",
-    "semver": "^7.6.3",
-    "ts-jest": "^29.2.4",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.5.4",
-    "ws": "^8.14.2"
+    "semver": "^7.6.3"
   },
   "dependencies": {
-    "@homebridge/plugin-ui-utils": "^1.0.3",
-    "axios": "^1.7.7",
+    "axios": "^1.7.8",
     "bottleneck": "2.19.5",
     "debug": "4.3.7",
     "hap-node-client": "^0.2.9",
     "is-my-json-valid": "2.20.6",
-    "mqtt": "5.10.1"
+    "mqtt": "5.10.3"
   },
   "author": "NorthernMan54",
   "license": "Apache-2.0",
@@ -62,26 +64,13 @@
   },
   "homepage": "https://github.com/NorthernMan54/homebridge-alexa#readme",
   "jest": {
-    "moduleFileExtensions": [
-      "js",
-      "json",
-      "ts"
-    ],
-    "rootDir": "./",
     "testEnvironment": "node",
-    "testRegex": ".test.ts$",
-    "testTimeout": 10000,
-    "transform": {
-      "^.+\\.(t)s$": "ts-jest"
-    },
-    "collectCoverageFrom": [
-      "**/*.ts"
+    "modulePathIgnorePatterns": [],
+    "coverageReporters": [
+      "lcov"
     ],
-    "coveragePathIgnorePatterns": [
-      "/ui/",
-      "/test/",
-      "/dist/",
-      "/scripts/"
+    "collectCoverageFrom": [
+      "src/**"
     ]
   }
 }
