+++ src/lib/alexa_smart_home_message_schema.json
