+++ src/lib/parse/Characteristic.js
