+++ test/hbConfig/.uix-dashboard.json
@@ -1,0 +1,1 @@
+[{"component":"UpdateCentreWidgetComponent","x":0,"y":6,"cols":4,"rows":6,"mobileOrder":10,"hidePort":true,"hideOnMobile":false,"draggable":true},{"component":"HomekitBridgesWidgetComponent","x":15,"y":0,"cols":5,"rows":12,"mobileOrder":35,"hideOnMobile":false,"draggable":true},{"component":"CpuWidgetComponent","x":4,"y":0,"cols":5,"rows":3,"mobileOrder":40,"hideOnMobile":false,"draggable":true,"refreshInterval":10,"historyItems":60},{"component":"MemoryWidgetComponent","x":4,"y":3,"cols":5,"rows":3,"mobileOrder":50,"hideOnMobile":false,"draggable":true,"refreshInterval":10,"historyItems":60},{"component":"NetworkWidgetComponent","x":4,"y":6,"cols":5,"rows":3,"mobileOrder":55,"hideOnMobile":false,"draggable":true,"refreshInterval":10,"historyItems":60,"networkInterface":"en0"},{"component":"UptimeWidgetComponent","x":4,"y":9,"cols":5,"rows":3,"mobileOrder":60,"hideOnMobile":false,"draggable":true},{"component":"SystemInfoWidgetComponent","x":9,"y":0,"cols":6,"rows":12,"mobileOrder":70,"hideOnMobile":false,"draggable":true},{"component":"HapQrcodeWidgetComponent","x":0,"y":0,"cols":4,"rows":6,"mobileOrder":100,"hideOnMobile":false,"draggable":true},{"component":"HomebridgeLogsWidgetComponent","x":0,"y":12,"cols":20,"rows":7,"mobileOrder":1000,"hideOnMobile":true,"draggable":true}]
