+++ test/hbConfig/auth.json
@@ -1,0 +1,10 @@
+[
+  {
+    "id": 1,
+    "username": "test",
+    "name": "test",
+    "hashedPassword": "df121e72b850a058bd68f3d05b1f94dc985821a3885b5da38e263654f29843f98bb6d8e594f6042ff871a424602a9bff50dffa97f258bdc1dca4c79e87bac20c",
+    "salt": "64085e70da64670349f042d4c3d3cac75b20233ffeb71db68399973f60da077d",
+    "admin": true
+  }
+]
