--- tsconfig.build.json
@@ -1,11 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "exclude": [
-    "node_modules",
-    "test",
-    "**/*spec.ts",
-    "ui",
-    "public",
-    "scripts"
-  ]
-}
