--- tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "declaration": true,
-    "removeComments": true,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "target": "es2020",
-    "sourceMap": true,
-    "outDir": "./dist",
-    "baseUrl": "./",
-  },
-  "exclude": [
-    "node_modules",
-    "./dist",
-    "ui",
-    "public"
-  ]
-}
